% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{clean_mrk}
\alias{clean_mrk}
\title{Clean genetic markers}
\usage{
clean_mrk(
  data,
  na_freq = NULL,
  rm_mono = FALSE,
  maf = NULL,
  recode = FALSE,
  rm_dupes = FALSE,
  keep = NULL,
  impute = FALSE,
  homo = as.character(c("A", "C", "G", "T")),
  hetero = as.character(c("R", "Y", "S", "W", "K", "M")),
  homo_num_min = -1L,
  homo_num_maj = 1L,
  hetero_num = 0L,
  row_names = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{data}{Either a \code{matrix} or a \code{data.frame} containing marker
genotype data. Markers must be arranged in rows and genotypes
(i.e. lines, hybrids, clones or individuals) in columns. Marker 
names can be provided as row names or in the first column (see 
argument \code{row_names}).}

\item{na_freq}{Upper limit for the frequency of missing values.}

\item{rm_mono}{\code{TRUE} to remove all monomorphic markers.}

\item{maf}{Lower limit for the minor allele frequency.}

\item{recode}{\code{TRUE} to recode to numeric code.}

\item{rm_dupes}{\code{TRUE} to remove duplicate markers. Markers sharing
identical information with missing values for different 
genotypes are also considered duplicates.}

\item{keep}{Character vector containing markers to keep.}

\item{impute}{\code{TRUE} to impute data with mean (for numeric data) or
mode (for categorical data).}

\item{homo}{Character vector defining homozygous marker genotypes. Default
is c("A", "C", "G", "T") with "A" standing for homozygous for 
nucleotide A.}

\item{hetero}{Character vector defining heterozygous marker genotypes.
Default is ("R", "Y", "S", "W", "K", "M") with "R" standing 
for a heterozygous genotype with nucleotides A and G according
to IUPAC code.}

\item{homo_num_min}{Integer value used for recoding of genotypes homozygous
for minor allele.}

\item{homo_num_maj}{Integer value used for recoding of genotypes homozygous
for major allele.}

\item{hetero_num}{Integer value used for recoding of heterozygous genotypes.}

\item{row_names}{Must be \code{TRUE} if marker names are provided as row
names.}

\item{verbose}{\code{TRUE} to print results at each step.}
}
\value{
\code{data.frame} with filtered marker genotypes.
}
\description{
\code{clean_mrk} filters and/or recodes genetic markers such as single
nucleotide polymorphism markers.
}
